setup the scripts to use TS
