+++ .eslintrc.js
@@ -1,10 +1,19 @@
 module.exports = {
   root: true,
   env: {
-    es6: true,
+    es2020: true,
     node: true,
   },
-  plugins: ['prettier'],
-  extends: ['eslint:recommended', 'plugin:prettier/recommended'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 2020,
+  },
+  plugins: ['@typescript-eslint', 'prettier'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
+    'prettier/@typescript-eslint',
+  ],
   ignorePatterns: ['abis/', 'build/', 'generated/', 'node_modules/', '!.prettierrc.js'],
 }
