+++ src/.eslintrc.js
@@ -1,15 +1,4 @@
 module.exports = {
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 2020,
-  },
-  plugins: ['@typescript-eslint', 'prettier'],
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:prettier/recommended',
-    'prettier/@typescript-eslint',
-  ],
   rules: {
     // NOTE: AssemblyScript doesn't have the same non-null type guard logic.
     '@typescript-eslint/no-non-null-assertion': 'off',
