+++ src/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./node_modules/assemblyscript/std/assembly.json",
+  "include": ["**/*.ts"]
+}
